Split into 3 files.
Restructure code to keep things clearer - work can now begin exclusively on the `MoleculeSet` file with the given capabilities of `Lattice` class.
